updated styling on profile routes, fixed save state of code in runner editor, added 2nd editor for editing existing code
