+++ src/App.js
@@ -3,7 +3,8 @@ import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import Login from "./Pages/Login/Login";
 import FileUpload from "./Pages/FileUpload/FileUpload";
 import Home from "./Pages/Home/Home";
-import Editor from "./Pages/Editor/Editor";
+import RunnerEditor from "./Pages/Editors/RunnerEditor";
+import ProfileCodeEditor from "./Pages/Editors/ProfileCodeEditor";
 import Peep from "./Pages/Peep/Peep";
 import Register from "./Pages/Register/Register.jsx";
 import Profile from "./Pages/Profile/Profile.jsx";
@@ -29,7 +30,8 @@ function App() {
           <Route path="/Login" exact component={Login} />
           <Route path="/Home" exact component={Home} />
           <Route path="/Profile" exact component={Profile} />
-          <Route path="/Editor" exact component={Editor} />
+          <Route path="/Editor" exact component={RunnerEditor} />
+          <Route path="/ProfileEditor" exact component={ProfileCodeEditor} />
           <Route path="/Peep" exact component={Peep} />
           <Route path="/Register" exact component={Register} />
         </Switch>
