+++ src/Styles/style.scss
@@ -183,21 +183,38 @@ box-shadow: 0px 12px 25px -4px rgba(0, 0, 0, 0.5);
     width:  5rem;
   }
 }
+.profLangPic{
+  height: 5rem;
+  width: 5rem;
+}
+@media only screen and (max-width: 800px) {
+  .profLangPic {
+    height: 3rem;
+    width: 3rem;
+  }
+}
+
+@media only screen and (max-width: 400px) {
+  .profLangPic {
+    height: 3rem;
+    width:  3rem;
+  }
+}
 .peepPic{
-  height: 8rem;
-  width: 8rem;
+  height: 6rem;
+  width: 6rem;
 }
 @media only screen and (max-width: 1200px) {
   .peepPic {
-    height: 5rem;
-    width: 5rem;
+    height: 4rem;
+    width: 4rem;
   }
 }
 
 @media only screen and (max-width: 800px) {
   .peepPic {
-    height: 4rem;
-    width: 4rem;
+    height: 3rem;
+    width: 3rem;
   }
 }
